+++ filelist.php
@@ -74,6 +74,7 @@ if (IS_GUEST && $torrent->isPrivate()) {
     bb_die($lang['BT_PRIVATE_TORRENT'], 403);
 }
 
+// Get torrent files
 $files = $torrent->$t_version_field()->$t_files_field();
 if ($meta_v1 && $meta_v2) {
     $files = new \RecursiveIteratorIterator($files); // Flatten the list
@@ -95,11 +96,26 @@ foreach ($files as $file) {
 $torrent_name = !empty($t_name = $torrent->getName()) ? str_short(htmlCHR($t_name), 200) : $lang['UNKNOWN'];
 $torrent_size = humn_size($row['size'], 2);
 
+// Get announcers list
+$announcers_list = $torrent->getAnnounceList()->toArray();
+$announcers_count = 0;
+foreach ($announcers_list as $announcer) {
+    $announcers_count++;
+    $row_class = ($announcers_count % 2) ? 'row1' : 'row2';
+    $template->assign_block_vars('announcers', [
+        'ROW_NUMBER' => $announcers_count,
+        'ROW_CLASS' => $row_class,
+        'ANNOUNCER' => $announcer[0]
+    ]);
+}
+
+// Output page
 $template->assign_vars([
     'PAGE_TITLE' => "$torrent_name (" . $torrent_size . ")",
     'FILES_COUNT' => sprintf($lang['BT_FLIST_FILE_PATH'], declension(iterator_count($files), 'files')),
     'TORRENT_CREATION_DATE' => (!empty($dt = $torrent->getCreationDate()) && is_numeric($creation_date = $dt->getTimestamp())) ? date('d-M-Y H:i (e)', $creation_date) : $lang['UNKNOWN'],
     'TORRENT_CLIENT' => !empty($creator = $torrent->getCreatedBy()) ? htmlCHR($creator) : $lang['UNKNOWN'],
+    'TORRENT_PRIVATE' => $torrent->isPrivate() ? $lang['YES'] : $lang['NO'],
 
     'BTMR_NOTICE' => sprintf($lang['BT_FLIST_BTMR_NOTICE'], 'https://github.com/kovalensky/tmrr'),
     'U_TOPIC' => TOPIC_URL . $topic_id,